delete user account and route to login screen
